+++ Commands/eco/pay.js
@@ -23,7 +23,7 @@ export default {
     // }
 
     if (payee.balance < amount) {
-      return interaction.reply({ content: 'You do not have enough money to pay that amount.', ephemeral: 'true' });
+      return interaction.reply({ content: 'You do not have enough money to pay that amount.', flags: MessageFlags.Ephemeral });
     }
 
     await ecoUtils.updateUser(payee.discordID, payee.balance - amount).then(async () => {