+++ Main/Events/Interactions/interactionCreate.js
@@ -66,12 +66,12 @@ export default {
       if(interaction.replied === false) {
         interaction.reply({
           content: 'An error occured whilst running that command. Please try running it again. If the error persists, please contact the bot owner.',
-          ephemeral: true
+          flags: MessageFlags.Ephemeral
         });
       } else if(interaction.replied === true) {
         interaction.editReply({
           content: 'An error occured whilst running that command. Please try running it again. If the error persists, please contact the bot owner.',
-          ephemeral: true
+          flags: MessageFlags.Ephemeral
         });
       }
     }